Singleplayer client/host iteration
Semi functional, there is something of a problem in that the client
turns the numbers it receives into hexidecimal.
However, apart from a few hiccups, it is operational.

The next step will be to remove the AI and replace it with a second
player by duplicating much of the programs logic for a second client.